Points and measures can now be deleted; selections working; multiple and global editing now supported
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@2928 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/qisis/apps/cneteditor/CnetTableColumnList.cpp
@@ -19,11 +19,11 @@ namespace Isis
   /**
    * @returns minX, maxX in a pair
    */
-  QPair<double, double> CnetTableColumnList::getVisibleXRange(
+  QPair<int, int> CnetTableColumnList::getVisibleXRange(
     int visibleColumn) const
   {
-    double minX = 0;
-    double maxX = 0;
+    int minX = 0;
+    int maxX = 0;
 
     QList<const CnetTableColumn *> visibleCols = getVisibleColumns();
 
@@ -38,7 +38,7 @@ namespace Isis
       maxX = minX + visibleCols.at(visibleColumn)->getWidth() - 1;
     }
 
-    return QPair<double, double>(minX, maxX);
+    return QPair<int, int>(minX, maxX);
   }
